WIP: SIG: SessPathPool turned into an interface
