+++ go/sig/egress/BUILD.bazel
@@ -2,7 +2,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["interface.go"],
+    srcs = [
+        "interface.go",
+        "sesspath.go",
+        "sesspathpool.go",
+    ],
     importpath = "github.com/scionproto/scion/go/sig/egress",
     visibility = ["//visibility:public"],
     deps = [
