+++ go/sig/egress/interface.go
@@ -16,7 +16,6 @@ package egress
 
 import (
 	"fmt"
-	"math"
 	"time"
 
 	"github.com/prometheus/client_golang/prometheus"
@@ -25,7 +24,6 @@ import (
 	"github.com/scionproto/scion/go/lib/common"
 	"github.com/scionproto/scion/go/lib/log"
 	"github.com/scionproto/scion/go/lib/ringbuf"
-	"github.com/scionproto/scion/go/lib/sciond"
 	"github.com/scionproto/scion/go/lib/snet"
 	"github.com/scionproto/scion/go/lib/spath/spathmeta"
 	"github.com/scionproto/scion/go/sig/mgmt"
@@ -99,104 +97,20 @@ type PathPool interface {
 	Destroy() error
 }
 
+type SessPathPool interface {
+	// Get best path.
+	Get(exclude spathmeta.PathKey) *SessPath
+	// Get path associated with the key.
+	GetByKey(exclude spathmeta.PathKey) *SessPath
+	// Use new set of paths.
+	Update(aps spathmeta.AppPathSet)
+	// Inform the path that the probe failed.
+	Fail(path *SessPath)
+	// Called periodically to make old failures less relevant.
+	ExpireFails()
+}
+
 type SessionSet map[mgmt.SessionType]Session
-
-const pathFailExpiration = 5 * time.Minute
-
-type SessPathPool map[spathmeta.PathKey]*SessPath
-
-// Return the most suitable path. Exclude a specific path, if possible.
-func (spp SessPathPool) Get(exclude spathmeta.PathKey) *SessPath {
-	var bestSessPath *SessPath
-	var minFail uint16 = math.MaxUint16
-	var bestNonExpiringSessPath *SessPath
-	var minNonExpiringFail uint16 = math.MaxUint16
-	for k, v := range spp {
-		if k == exclude {
-			continue
-		}
-		if v.failCount < minFail {
-			bestSessPath = v
-			minFail = v.failCount
-		}
-		if v.failCount < minNonExpiringFail && !v.IsCloseToExpiry() {
-			bestNonExpiringSessPath = v
-			minNonExpiringFail = v.failCount
-		}
-	}
-	// Return a non-expiring path with least failures.
-	if bestNonExpiringSessPath != nil {
-		return bestNonExpiringSessPath
-	}
-	// If not possible, return the best path that's close to expiry.
-	if bestSessPath != nil {
-		return bestSessPath
-	}
-	// In the worst case return the excluded path. Given that the caller asked to exclude it
-	// it's probably non-functional, but it's the only option we have.
-	return spp[exclude]
-}
-
-func (spp SessPathPool) Update(aps spathmeta.AppPathSet) {
-	// Remove any old entries that aren't present in the update.
-	for key := range spp {
-		if _, ok := aps[key]; !ok {
-			delete(spp, key)
-		}
-	}
-	for key, ap := range aps {
-		e, ok := spp[key]
-		if !ok {
-			// This is a new path, add an entry.
-			spp[key] = NewSessPath(key, ap.Entry)
-		} else {
-			// This path already exists, update it.
-			e.pathEntry = ap.Entry
-		}
-	}
-}
-
-// A SessPath contains a path and metadata related to path health.
-type SessPath struct {
-	key       spathmeta.PathKey
-	pathEntry *sciond.PathReplyEntry
-	lastFail  time.Time
-	failCount uint16
-}
-
-func NewSessPath(key spathmeta.PathKey, pathEntry *sciond.PathReplyEntry) *SessPath {
-	return &SessPath{key: key, pathEntry: pathEntry, lastFail: time.Now()}
-}
-
-func (sp *SessPath) Key() spathmeta.PathKey {
-	return sp.key
-}
-
-func (sp *SessPath) PathEntry() *sciond.PathReplyEntry {
-	return sp.pathEntry
-}
-
-func (sp *SessPath) IsCloseToExpiry() bool {
-	return sp.PathEntry().Path.Expiry().Before(time.Now().Add(SafetyInterval))
-}
-
-func (sp *SessPath) Fail() {
-	sp.lastFail = time.Now()
-	if sp.failCount < math.MaxInt16 {
-		sp.failCount += 1
-	}
-}
-
-func (sp *SessPath) ExpireFails() {
-	if time.Since(sp.lastFail) > pathFailExpiration {
-		sp.failCount /= 2
-	}
-}
-
-func (sp *SessPath) String() string {
-	return fmt.Sprintf("Key: %s %s lastFail: %s failCount: %d", sp.key,
-		sp.pathEntry.Path, sp.lastFail, sp.failCount)
-}
 
 type SessionSelector interface {
 	ChooseSess(b common.RawBytes) Session
