+++ go/sig/egress/sesspath.go
@@ -1,0 +1,52 @@
+// Copyright 2018 ETH Zurich
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package egress
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/scionproto/scion/go/lib/sciond"
+	"github.com/scionproto/scion/go/lib/spath/spathmeta"
+)
+
+// A SessPath contains a path and metadata related to path health.
+type SessPath struct {
+	key       spathmeta.PathKey
+	pathEntry *sciond.PathReplyEntry
+	lastFail  time.Time
+	failCount uint16
+}
+
+func NewSessPath(key spathmeta.PathKey, pathEntry *sciond.PathReplyEntry) *SessPath {
+	return &SessPath{key: key, pathEntry: pathEntry, lastFail: time.Now()}
+}
+
+func (sp *SessPath) Key() spathmeta.PathKey {
+	return sp.key
+}
+
+func (sp *SessPath) PathEntry() *sciond.PathReplyEntry {
+	return sp.pathEntry
+}
+
+func (sp *SessPath) IsCloseToExpiry() bool {
+	return sp.PathEntry().Path.Expiry().Before(time.Now().Add(SafetyInterval))
+}
+
+func (sp *SessPath) String() string {
+	return fmt.Sprintf("Key: %s %s lastFail: %s failCount: %d", sp.key,
+		sp.pathEntry.Path, sp.lastFail, sp.failCount)
+}
